requirements.txt
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
tavily-python
langgraph-cli
python-dotenv
langchain
pydantic
PyGithub
gitpython 
docling
langchain-chroma
langchain-docling
ipywidgets
langgraph-cli[inmem]
langchain_qdrant